refactor: re-utilize syncWithGatsby in handleDeltaChanges
